tests: Map WORKDIR to W: to shorten paths
In some cases Wine (and Windows proper) struggle with very long paths;
in particular the way that gcc invokes the 'cc1' subprogram is limited
in how deep the path to the subprogram can be. This can cause issues
when testing the SDK under wine, as the paths can easily get quite long
and exceed this limit. In order to work around this, setup the Wine test
context so that the W: drive maps to the SDK image ${WORKDIR}, which
allows wine to effectively use paths relative to this directory making
them significantly shorter.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>
(cherry picked from commit f40731611e9682928db78c83db72d384c2edee10)