oeqa: Set LANG in wine tests
Glib likes to try and determine the local name using the Windows
GetConsoleOutputCP(), which doesn't work well running under wine in our
test cases (it seems to be linked to stdin not being a terminal).

This results in the error message:

GLib-WARNING (recursed) **: Failed to determine console output code page: Invalid access.. Falling back to UTF-8

and then Glib falls back to the locale "UTF-8", which is invalid and the
program will likely crash.

In particular, pkg-config seems to be sensitive to this on some hosts
(e.g. Ubuntu 24.04)

Set the LANG variable to a valid language to make sure that this
codepath in Glib is not triggered.

Signed-off-by: Joshua Watt <JPEWhacker@gmail.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>