libiconv: tell autoreconf where to find supplementary macros
The libiconv build is non-idiomatic, and most importantly doesn't use
automake so there's no way to tell aclocal where to find macros.

Currently that isn't an issue because our autotools class explicitly
specifies any directories it can find containing macros but that will
be removed shortly, so tell autoreconf where to find the macros as the
upstream Makefile.devel does.

Note that we need to pass absolute paths here because libiconv has a
nested configure script which fails when autoreconf is used with
relative paths.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>