core-image-mingw-sdktest: depend on nativedk-dbus-tools, not -dbus
The dbus package is the bus daemon itself, which doesn't make a lot of
sense in the context of the host component of a SDK. From "dbus: add
message-bus PACKAGECONFIG" in oe-core, nativesdk builds of dbus won't
build the bus at all.

Instead, change the dependency in this recipe to nativedk-dbus-tools,
which are marginally more useful in a SDK context and still exercise
the building of the recipe.

Signed-off-by: Ross Burton <ross.burton@arm.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>